Auth flow added - refactors going well
