+++ src/schemas/openai.types.ts
